requirements.txt
streamlit
google.geminiai
python-dotenv